Separate some reusable integration helpers
